Add rcodesign installation to golang-builder
Signed-off-by: Alper Polat <gitperr@gmail.com>

Install rust and rcodesign

This method was chosen because rust shipped with debian is 1.63.0,
which does not seem to be sufficient for the binary to be built.
So, we have to install rust from source.

Signed-off-by: Alper Polat <gitperr@gmail.com>

Move rust installation further down for better flow

Seems like build pipeline fails because we remove the osxcross
directory, so might be better to concatenate the path and install
rust a bit later.

Signed-off-by: Alper Polat <gitperr@gmail.com>

Use rust image to add rcodesign instead of rustup

Signed-off-by: Alper Polat <gitperr@gmail.com>

Copy from the correct directory

Seems like the rcodesign binary ended up in `/usr/local/cargo`
not /root...
Signed-off-by: Alper Polat <gitperr@gmail.com>

Use `/usr/local/bin` instead

Keep it simpler, we don't need the `cargo` directory.
Signed-off-by: Alper Polat <gitperr@gmail.com>

Remove redundant space

Signed-off-by: Alper Polat <gitperr@gmail.com>